requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
oracledb==3.4.1
pillow==12.1.0
pycparser==2.23
python-dotenv==1.2.1
reportlab==4.4.7
typing_extensions==4.15.0
